older version of kmk provided in example
Yhe kmk firmware repository provided in this project requires a version of the kmk firmware from 2022 and this is not specified anywhere in the example repository and it took me a solid hour to figure this out, maybe this will help out other people to find the right version of kmk. The creator of that repo did not update the readme file (i belive he is no longer active), so i figured this was the only other place i could specify the required firmware.